cloud_functions_codes/pda-02-start-job-transfer/requirements.txt
functions-framework==3.*
google-cloud-storage-transfer
google-auth
google-cloud-bigquery
google-cloud-storage
openpyxl
pytz
protobuf